Quite a bunch of modifications here.  I removed a couple of the src rpms from our srpms directory, since they were coming directly from rawhide.  Instead, I have now set up a yum repo on my people page with a bunch of the different packages we are using.  So we still use an F-8 base, but with updated packages. Along with this, I updated the ovirt.ks host builder kickstart to point there.
I also updated the collectd RPM to fix a bug where it wouldn't build on i386.

The kvm package was updated to fix a bug when building against a kernel that
isn't running; I was basically missing a configure argument.

The invirt-wui spec file was updated with a bunch more Requires, specifically
for ruby-kerberos, the version of rails, and the version of ruby-libvirt.

I updated the host-browser code to put new machines into the default hardware
resource group.

Finally, a bunch of updates to the taskomatic.  First, moved over to the updated
ruby-libvirt binding names (no more StudlyCaps).  Second, updated some failure
messages to be more informative.  Third, updated the disk building section so
we can handle more than 1 disk at a time (the first one is always the boot
device).  Finally, made start_vm respect the hardware_resource_groups when
choosing a host to start this VM on.