A few modifications here.  First, remove the ovirt-init and ovirt-install.py files.  The init script is no longer used since we now use rjones livecd-to-pxe script.  ovirt-install.py was never used, but was just an example of what would be needed to make virt-install work with the remote stuff.
Next, update common-post.ks to skip a 1 second sleep if it was successful
on the first pass (the common case).

Also make sure to remove /etc/krb5.conf, and then fetch it on bootup.  Also
remove the check for "ovirtbr0", since that is just a bad assumption.  Instead,
as we go through the interfaces, we try each one until we successfully get a
krb5.tab and a krb5.conf.