some cleanup of ovirt-host kickstart to make images smaller
This patch adds --excludedocs to the kickstart for the x86_64 and i386
ovirt host kickstarts.  For the most part we're removing docs manually
right now anyhow, but this may catch some other things.

Also cleaned up the section that removes docs and localization.  Recreate
the cracklib dicts (frees up about 8MB) and remove unneeded kernel
modules.  Most of this stuff will just end up in the whitelist/blacklist
module, but the cracklib stuff is a different issue.  They can't be
removed, they need to be recreated.  This will probably need to continue
to happen in %post.

The end result of this cleanup is that the image boots and it is 58MB
compressed/162MB uncompressed (filesystem is 450MB total)

One of the things we need to fix is the filesystem total size.  The
filesystem doesn't need to be 450MB.  It could be initially created as
450MB and then resized to 200MB or less.

Signed-off-by: Perry Myers <pmyers@redhat.com>
Acked-by: Chris Lalancette <clalance@redhat.com>