fix modification of ovirt-wui.conf
ovirt-wui.conf modification in the create-default-principals script was
not occurring because of an errant if statement.  This has been removed.
In addition, the file that comes in the RPM already has the virtual host
and namevirtual host directives, so these do not need to be in the script.
   The modification of the Krb5KeyTab can also be removed since the
default value in the rpm file is correct.

Signed-off-by: Perry Myers <pmyers@redhat.com>
Acked-by: Chris Lalancette <clalance@redhat.com>