Restructured the wui production and devel kickstart posts so that most of the functionality is moved into two scripts (ovirt-add-host and ovirt-wui-install).
The post sections now contain just those steps that are specific to creating
a production or development appliance.  This way, users who want to install
on existing OSs can just run ovirt-wui-install manually.

Spec file is updated to pull in the new scripts and omit the outdated scripts.
Removed the references to the FreeIPA F7 repo, since those RPMs were very
out of date.  Until we move to F9, we need to put the ipa RPMs into the
ovirt-management repository.

NOTE: The ovirt-fix-ipa script is a temporary hack so that the ipa server
can run on the same host as the oVirt WUI.  As soon as we get RPMs from
FreeIPA that have the ipa server running in /ipa this will go away.

Signed-off-by: Perry Myers <pmyers@redhat.com>
Acked-by: Chris Lalancette <clalance@redhat.com>