This refactors the Hardware Pool and VM Resource Pool models. The OrgPool/NetworkMap/HostCollection subclasses of HardwarePool are removed entirely -- the HardwarePool hierarchy is once again completely flexible. You can have a single host pool or a deep hierarchy. VM Library has  been renamed to VM Resource Pool. At a model level, Hardware Pools and VM Resource Pools now inherit from a common (abstract) Pool class -- but this is not exposed in the UI.
In addition to the simpler (and more flexible) Hardware Pool hierarchy, the other main focus for this change is to allow a hierarchy of VM Resource Pools. The model now supports this, although the views do not (yet) allow the creation of nested VM Resource Pools. In addition, the quota/resource usage calculations do not yet deal with nested VM Pools.