Some more refactoring here.  The important point to note here is that we are now splitting up our releases into 3 different categories:
1)  Developer - an appliance that has all services running, including DHCP,
DNS, TFTP, ovirt-wui.  Has a private network, and runs on the same box as
the "fake" managed nodes.

2)  Bundled - an appliance that has all services running, including DHCP,
DNS, TFTP, ovirt-wui.  Is bridged over to the real network, so can manage
other hosts cabled to the same network.  Requires a separately cabled network.

3)  Essential - an appliance that only runs the ovirt-wui services.  This is
bridged over to a real network, and depends on the real network for DHCP and
other services.  This does not require a separately cabled network.

This commit makes it so that developer and bundled all use the same kickstart
file, which removes a lot of duplication and makes things much simpler.