allow create-wui-appliance to use local repo tree along with ISO
This patch modifies create-wui-appliance.sh so that both a local ISO or an
install tree can be used to install the WUI appliance.  The advantage of
using a local tree is that a kickstart file can be specified so that
manually entering the kickstart location during the boot process is not
necessary.

Signed-off-by: Perry Myers <pmyers@redhat.com>
Acked-by: Hugh Brock <hbrock@redhat.com>