Problem: ovirt-identify-node was calling libvirt functions.  However, libvirtd isn't running at this point in boot, so calling these functions fails.
Solution: Luckily, we can get almost all of this information from python's
built-in facilities (and we will be able to get it from bash later as well).  So
make native python calls to get this information.

Longer term, I think we'll probably want to do this in 2 stages (1st stage just
grab the keytab, and 2nd stage actually send over relevant information).  But
this should fix it for now.

Warning: it's only been lightly tested so far, so beware.

Signed-off-by: Chris Lalancette <clalance@redhat.com>

Acked-by: Hugh O. Brock <hbrock@harpcolumn.com>