Fix problem where start_vm fails if storage pool is not already active
When a vm is started the existing storage pools on the managed node are
scanned to see if the storage volume that the guest needs is present.
If it is present it is used, if not present it is created and started.

However, there is another case where the pool could be present and
inactive.  In this case (not sure how it gets in this state, but it
could happen) we need to start the pool.

So this code runs if the pool is defined and inactive, it activates

This fixes some bugs seen where the start action in the WUI would
fail silently.

Signed-off-by: Perry Myers <pmyers@redhat.com>