Add host state field
This patch adds a new column to the hosts table named 'state' which is
updated via host-status to "available" or "unavailable" based on whether
it is reachable.  I'm hoping the UI folks can incorporate this fairly
easily.

It also quiets down host-status and makes it so it only kicks taskomatic
when the state of a vm needs changing.

Signed-off-by: Ian Main <imain@redhat.com>
Acked-by: Scott Seago <sseago@redhat.com>