Cleanup the ovirt-wui SPEC and activate daemons
While I develop, I constantly do something like:

rpm -e ovirt-wui ; make rpms ; rpm -ivh rpm-build/ovirt-wui-<version>.rpm

However, this sequence does not currently turn on the various daemons.  The
attached patch always adds the relevant daemons in question, and turns them on
if and only this is a new install.  If it's an upgrade (and the daemons are
already in place), it respects the administrator's decision.

Signed-off-by: Chris Lalancette <clalance@redhat.com>