Errors starting the ovirt-host-collect daemon
When I attempted to start the ovirt-host-collect daemon, it failed with a
backtrace saying there was no unixsock present.  Which, indeed, is because we
don't have the /etc/collectd.conf setup properly on the WUI appliance.  After
adding that configuration, it still hung up on start because "do_daemon" was set
to false by default.  The attached patch fixes both of these issues and allows
the ovirt-host-collect daemon to start.

Signed-off-by: Chris Lalancette <clalance@redhat.com>