Taskomatic exception cleanup
Attached is a rather large patch that cleans up taskomatic exception handling.
It's quite nice in that it removes a lot of code, and moves almost all exception
handling code to the top-level, where it really belongs.  This will also give us
backtraces and good exception messages when things fail.  The only real downside
to this patch is that I've only lightly tested it so far, and it is a pretty big
patch for right now.  If people are very against this patch, I can probably come
up with something smaller (but not as nice) for the time being.  Comments?

Signed-off-by: Chris Lalancette <clalance@redhat.com>