console plugin integration, first pass.
I added a new attribute to vms, vnc_port that needs to be filled in by taskomatic when starting/resuming VMs. If vnc_port (and host) are set for a VM, an 'open console' link appears in the VM details pane which brings up a (currently 800x600) popup window for the VM console (using the x-gtk-vnc plugin).

For this to actually work, we need the following:
1) taskomatic changes to fill in vnc_port on start (and null it out when setting state to stopped, etc)
2) x-gtk-vnc firefox plugin needs to be installed on whatever box you're using to browse

EDIT: also make sure VM is running before producing console link.
EDIT 2: pull out "has console" test into model so that the controller can also verify in the console handler