Add vnc port to database.
This patch parses the xml from the newly created vm to get the vnc port
and adds it to the database.  If the host is stopped the port is set
to nil.  This works for stopping of vms including the case where the
node reboots because taskomatic is notified through host-status.

There is also some blanks removed and an indentation fixup.