Use multiple processes to check host status
This patch causes host-status to fork() up to node_count/5 times to
connect out to hosts via libvirt.  This guarantees that that it takes at
most 5 timeouts in a row to verify all nodes.  This should help with the
bottleneck we were seeing with libvirt connect timeouts.  Testing with 105
nodes, almost all of which were down, it took 27s to query all of them.

Signed-off-by: Ian Main <imain@redhat.com>
Acked-by: Hugh O. Brock <hbrock@redhat.com>