Fixed gitbump so that it is always greater than the last release
gitbump would always use a version release with the release starting
at 0.

This is bad when the version hosted on ovirt.org is release 1 or higher
So gitbump is modified to always use the release number in the version file
and then increment the decimal portion.

i.e. 0.90 1 becomes 0.90 1.1 instead of 0.90 0.1

Signed-off-by: Perry Myers <pmyers@redhat.com>