consolidated the storage action links.
Add and Create Storage Pool are now two tabs on the same popup widget. I've also consolidated delete and remove, adding a confirmation dialog for the user to choose which action to take (as specified in the design mockups).

One thing remaining for a follow-on task is to add some css styling for the  radio buttons and text to make it conform better to the design mockups.