Merge developer/bundled appliance into a single appliance: ovirt-appliance
The separation of the ovirt appliance image into developer/bundled versions
is not necessary, since the only differences between them are in how the
host bridge is configured.  Developer assumes that you only have a single
NIC on your host so you can't have a separate ovirt network managed by
the appliance.  Bundled assumes you have the second NIC available.

These patches remove the separation between bundled/developer.  The new
appliance is called ovirt-appliance and can be installed with or without
a bridge to a real network device.  In either case dummy managed nodes
are allowed (they can be used in conjunction with the real managed nodes
or just ignored in that case)

Both build-all.sh and create-wui-appliance.sh have the developer/bundled
flags removed, and replaced with -e flag to indicate the ethernet device
to use as the bridge.  If this is omitted it is assumed that the ovirtbr
will not bridge to a physical network.  The bridge (if to a physical net)
is made persistent in rc.local (eventually libvirt should handle this)
If the device to use on the bridge is used already, the script aborts
(we don't want to mess up someones eth0 config by accident)

Because the appliance name, bridge name and appliance image name have
changed (ovirt-appliance, ovirtbr, ovirt-appliance.img) there are
sections of the script that clean up the older named appliances.

NOTE: This revised patch incorporates suggestions from Chris L. and Jeff
Schroeder.  Specifically fixing some of the checks for the presence of
$bridge, making rc.local executable and fixing the usage syntax for
build-all.sh

Signed-off-by: Perry Myers <pmyers@redhat.com>