Move building of PXE image into RPM build step
Old ovirt-host-image rpm used to store a copy of the ISO and the
tftpboot directory.  This is redundant.  Source file for RPM is now
just the ISO and build step uses ovirt-pxe script to build the tftpboot
directory from the ISO

ovirt-pxe, ovirt-flash and ovirt-flash-static are now all part
of the ISO RPM so that users can easily deploy the ISO image onto
these media types.

Signed-off-by: Perry Myers <pmyers@redhat.com>