permission denormalization. (revised)
Permission records are now stored for all pools a user has access to, even implied/inherited permissions. This means that if an admin is granted access to the 'default' pool, there will be one permission record for this pool, and an additional "inherited" record for each pool below 'default' in the hierarchy -- for the rest of these records, inherited_from_id will point to the master permission record.

Any time a user is granted permission on a pool, the inherited records are automatically created. When a new pool is created, inherited records are created here too.

You need to run 'rake db:migrate' to get the updated permissions table --  the migration script will take care of creating inherited records for existing permissions.

In the UI, both direct and inherited grants are shown, but only direct grants have the checkbox to allow modification/deletion.