Add serial console support
This patch adds serial console support to the ovirt node.
With this patch, nodes in developer mode now support a serial console
and can be connected to with the virsh console command.

The speed is set to 115200 by default but is configurable
via the kickstart option.

This requires the new kvm and livecd-creator from the ovirt
repo in order to work properly!

Signed-off-by: Ian Main <imain@redhat.com>
Acked-by: Alan Pevec <apevec@redhat.com>