Managed node configuration generator.
This generator takes as input an object graph for a Host and a map of mac
addresses to interface names on the managed node. It then generates the
configuration details for that managed node (currently, only the NIC
configuration is processed). With those details it generates the content
for a configuration file that will be retrieved by the managed node and
run through augtool to configure the managed node.