Fix how taskomatic adds NICs to the System prior to creating it in Cobbler.
Taskomatic was using the old constructor for Cobbler::NetworkInterface, which took
two arguments: the interface name and a map of the attributes. With rubygem-cobbler
v0.0.2 that constructor was changed to take just the map of attributes, and gets
the interface name from there.

This patch fixes the call to create the NetworkInterface when creating a new VM,
by removing the interface name.

Signed-off-by: Darryl L. Pierce <dpierce@redhat.com>