Added Tasks tab/pane to HW and VM pools pages
A tasks pane is now available on both VM and HW pools page. In both cases, the  tasks show up as a flexigrid panel showing tasks for the pool. By default, tasks are filtered for 'pending' state, but they may be filtered on any state (or to show all). For the VM page, all tasks for all VMs in the VM pool are available. For the HW pool page, there are 3 types of  tasks available:
1) VM tasks: tasks for all VMs contained in VM pools within this HW pool
2) Host tasks: tasks for all hosts in this pool
3) storage tasks: tasks for all storage pools in this HW pool.
By default all task  types are shown, but the pulldown menu lets the user filter on task type as well.