The node now passes in the mac address and iface names during identify.
This works with the patch for the server the generates a configuration file
for the node based on the interface names passed in with each mac address.

Additionally, the node will pull down the configuration and process it as a
script. Since the configuration will be composed of two separate parts:

1. a script to load one or more kernel modules
2. a configuration file to pass to augtool

The script is downloaded to a temp directory and then passed to bash. If
either of the two expected files is produced then it is further processed to
configure the node.