Check to make sure a config file was applied, if not revert to default config
ovirt-early tries to use the managed_node controller on the server
to get the remote config.  If there is none available a file is returned
but there is presently no way to determine if that file is 'empty' or not

This check looks to see if any ifup-eth* files were created.  If none were, then
we assume the config was blank and use the default bridge config of one bridge
per physical interface.

Also did some refactoring of log output for the configure network interfaces
function so that it's easier to see what it going on for debugging purposes.

Signed-off-by: Perry Myers <pmyers@redhat.com>