add ovirt semodule in Node
For now, it is only to allow qemu to access disk partitions directly,
required in order to use iSCSI storage pools with SELinux enabled.

Signed-off-by: Alan Pevec <apevec@redhat.com>

Moved from ovirt-node-image repository as it should be in the node
RPM since that RPM is used for creating nodes from stock Fedora
installs and this policy needs to be set there as well.  Added check
for selinuxenabled before making the change.

This is necessary to make clalance's patches for allowing the appliance
to manage the host it is running on as a Node.

FIXME: This patch is going in to fix the problem, but we should be using
http://fedoraproject.org/wiki/PackagingDrafts/SELinux/PolicyModules

Signed-off-by: Perry Myers <pmyers@redhat.com>