Updated the configuration processing to remove the need for scripts.
The system now takes an encoded configuration descriptor from the server. It
then parses from that a set of aliases for bondings if such exist. It then
also extracts configuration details for the various network interfaces on the
node. Afterward, it reloads module dependencies and then restarts the
networking service.

This patch also includes updated rdocs for the configuration generator.