Added support for local storage configuration.
NOTE: This is a reduxed patch based on Jim Meyering's changes, and includes
updates to support automated execution.

This patch wipes out the entire disk on the node and replaces it with
a set of partitions. The partitions are a raw partition to hold the bootable
kernel, and the rest of the disk is taken up as an physical volume.

On the physical volume is created the following logical volumes:

 * SWAP    - swap partition
 * ROOT    - to host the node image
 * CONFIG  - to hold node configuration data
 * LOGGING - to hold local logs
 * DATA    - to hold VM images