Enabled automated network configuration via kernel arguments.
To configure IPv6, the OVIRT_IPV6 variable will have one of three
values:

"dhcp" - Configure stateful automatic addressing,
"auto" - Configure stateless automatic addressing,

or it can contain an IPv6 address and manual configuration will be
performed. If the argument is not present then IPv6 is disabled.

IPv4 addressing is assumed, and DHCP will be used, unless the
OVIRT_IP_ADDRESS variable contains the value "off". If it contains
any other value, then that is assumed to be  an IPv4 address.

OVIRT_IP_NETMASK and OVIRT_IP_GATEWAY are option and will be used
if present and a static IP address were being defined.