Created stateful and stateless subpackages used for installing on bare metal Fedora hosts
ovirt-node-stateless is for use on the embedded oVirt Node.
ovirt-node-stateful is for use on already installed Fedora hosts to configure
that host to act as a Node.  Presently it only works when run on the same
host as the oVirt Appliance.

ovirt-node RPM provides common binaries/scripts used by the subpackages

Splitting up the RPMs is required since ovirt-node RPM now packages up
gptsync which conflicts with anaconda.

Signed-off-by: Perry Myers <pmyers@redhat.com>