Fix mount_config to not abort if /config is already mounted
Need to do this because mount_config does two things
1. mounts /config
2. bind mounts all config files

And there is use case where /config is already mounted but there
are additional files to be bind mounted in /config

This fixes a problem with readonly root mounting of config files in the
default init scripts.  readonly root in init requires files to be present
for a bind mount to work (i.e. it doesn't do a touch first).  In the cases
of sysconfig/network-scripts/ifcfg-* these files will not necessarily be
present during Node creation.  Therefore we can't rely on readonly root
to bind mount these files and need to rely on our mount_config function
instead.

Signed-off-by: Perry Myers <pmyers@redhat.com>