Reworked logging for standalone mode
Entire firstboot is logged to both console/logfile using tee.  Selected
sections of the code are set to only log stdout/stderr to logfile to prevent
data overload on console.  In these cases log() function is used to write
to the console.
TEE and log() must not be redefined in the scripts.

Signed-off-by: Perry Myers <pmyers@redhat.com>