Use --no-check-certificate with wget
This patch adds the --no-check-certificate to wget for all wgets since
the server now requires ssl and doesn't have a valid cert.  I'm not
certain this is the best way to handle this case but I'm posting this
patch anyway to see if it works and if it's an acceptable solution.
Note that I have not yet tested this either. :)

Signed-off-by: Ian Main <imain@redhat.com>
Acked-by: Jason Guiditta <jguiditt@redhat.com>