Report on physical network devices when identifying the node.
The default interface configuration is only applied if a configuration
was not retrieved from the server for some reason. It is only applied to
the management interface.

When ovirt-identify-node collects NIC details to send back to the
server, it only collects for physical devices that support the net.80203
capability as reported by HAL.