Fixes when more than one UDI is returned while getting a drive's size.
If more than one UDI is returned, they are delimited by a carriage
return. That is changed to a space, then the lot is iterated over. Each
UDI is tested to see if it has the property "block.is_volume=false". If
so then that UDI is for the underlying block device and not a logical
volume on the device.

Signed-off-by: Darryl L. Pierce <dpierce@redhat.com>