Make all yes/no prompts consistent. rhbz#508778
Added a new function, "ask_yes_or_no", to ovirt-functions. This method
allows the caller to indicate whether a choice is required (defaults to
true) and whether to allow the user to abort the question (defaults to
false).

Changed all prompts that ask for yes, no and/or abort to use this new
method.