Change DNS setup to match NTP. rhbz#508677
Rather than expecting the DNS entries to be on one line, they are now
entered as two separate entries.

Also fixes a bug that caused the first DNS entry to be lost when a
second entry was provided.

In both cases, the input is collected in a colon-delimited string that
is then processed. This is to provide consistency between the
interactive setup and the automated setup.

Signed-off-by: Darryl L. Pierce <dpierce@redhat.com>