Introduces the node-admin toolset.
Defines a primary entry point that displays a menu of options that user
can use to administer the managed node. It leverages the code from
virtinst-python to do the heavy lifting and just provides a front end
for collecting user data.

Created a new configuration class to be used for all configuration
screens.

The user can select to create a node. They are then walked through the
steps to create a domain similar to virt-manager.

The user can create a defined domain.

The user can destroy a created domain.

The user can undefine a defined domain.

The user can list all domains on the system.

The user can create a new user account.

When the RPM is created, a separate entry point is created for each
node admin command. This allows the commands to be invoked individually
as well as from the main menu.