Moves all ovirt-config-* scripts to /usr/libexec.
This is to reinforce with the user that the scripts are not meant to be
used at the commandline individually.

The spec file and Makefile are both fixed to change the installation
location. The ovirt-functions file now contains a variable which points
to the installation directory, which is used inside of all scripts which
call out to other scripts.

Updates the make system to use makefiles and to install the files there.