Fixes persisting a persisted file, or unpersisting a non-persisted file.
Previously the ovirt_store_config function only checked if a directory
was already persisted. This patch adds in checking if a targeted file
is already persisted as well. If so, then it fails with an error message.

When the user tries to unpersist a file that was not explicitly persisted
before (it's not listed in /config/files) then the user is told the file
was not persisted previously, and the command exits.

Resolves: rhbz#576329

Signed-off-by: Darryl L. Pierce <dpierce@redhat.com>