fix rhn timeout
We were running with a timeout of 90 seconds and the default
number of tries is 20 which resulted in an excessive wait time.
This patch changes it to 3 tries and 30 seconds per try.

Also, due to a bug in wget (rhbz#670241), a timeout does not result in an error.
The patch also works around the error by checking the downloaded
file size

rhbz#664378

Signed-off-by: Mike Burns <mburns@redhat.com>