Cleanup installation kargs
add reinstall alias for firstboot

add uninstall karg
if uninstall specified, wipe disk(s) mentioned in storage_init
in the dracut plugin, then reboot the machine

add install option as default option in installer
install option causes the gui installer to run instead of the admin
console.  This is set as the default when running from a
non-installed system.

rhbz#640782