store hostname in /etc/hosts if set manually
Fixes a performance issue that occurs when the hostname is not
resolvable and a configured dns server is unavailable.  All sudo
commands run in the above situation take 30 seconds to complete.

This is because of 6 interations of 5 second timeouts trying to get
to the broken dns server.

The fix is to set the hostname directly in /etc/hosts on the
127.0.0.1 line to ensure that sudo calls work correctly.

rhbz#768491

Change-Id: I0f9618b432ad592adc41db47605fb2854b7dd860