Detect whether fakeraid is enabled on installation disks
Using the dmraid tool, detect if a device selected for installation
(in either auto-install or the TUI) contains fakeraid metadata.

A new commandline option, wipe_fakeraid, has been added.  If this exists,
then all fakeraid metadata will be wiped from the selected disks.

If the wipe_fakeraid option is passed, a tui install will silently clear
all fakeraid data from the device prior to installation.  An auto-install
will also silently remove all fakeraid data from the selected devices.

If it is not passed, or is passed with a value of 0 or no, then an
auto-install will fail.  A TUI install will prompt the user if they
want to wipe fakeraid from the devices they've selected for installation.

Change-Id: Ibb8e01804ae91ff7f92c019ecbd7236659a3051d