early: Mount all entries in fstab.
When systemd get's started it reads the contents of the ro-rootfs
/etc/fstab file which is very likely different from the persisted
version and thus only the mountpoints given in the ro-vesion are
mounted. All entries appearing later (e.g. when the persisted file
is bind-mounted) are discovered by systemd, but not mounted
automatically.
To hack around this, a simple "mount -a" is run just before the
ovirt logic jumps in.

oVirt bug rhbz#806349
Upstream systemd bug rhbz#807203

Change-Id: I0713eaaa5074f2d24b7565db06686caaedc08c7a
Signed-off-by: Fabian Deutsch <fabian.deutsch@gmx.de>