show serial in description for sas/usb disks
Previously, we were always using ID_SCSI_COMPAT for the device
description in the storage selection screen.  If a device did
not have that defined, we would check if it was a virtio disk.
If it wasn't, we would display "unknown".  Now, we will check
the ID_SERIAL field and use that if ID_SCSI_COMPAT is not
available.  If both are unavailable, we will fall back to
"unknown"

rhbz#824854

Change-Id: I0732c9db59274e68f1c364cd9d83aa899d87eb42
Signed-off-by: Mike Burns <mburns@redhat.com>