installation: Copy existing log files
This patch fixes the copying of the log files from tmpfs to
HostVG-Logging near the end of the installation.
Because this happens only near the end of the installation not all
log entries logged during the installation to (especially) ovirt.log
are copied. But most files and relevant directories are copied.

rhbz#829283

Change-Id: I824188dcf0ae071e5b8b28e3e837548d6dcc8cfb
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>