fix check for kvm support
Previously, we used virsh to determine if kvm support existed.  We
wanted to remove the dependence on libvirt in the check, so we
tried using systemd-detect-virt, but that determines if you're
running in a vm and not if you're capable of running a vm.

Now, check for existence of /dev/kvm and for the kvm vendor module
to be loaded.  This will indicate if KVM is enabled.

rhbz#832223

Change-Id: Ia4656b82639999e0328d63a37cef09097b287bdb
Signed-off-by: Mike Burns <mburns@redhat.com>