network: Persist and load net rules
Previously the udev network rules were generated at each boot and thus
the NIC order was not persistent.
Now we are bind-mounting a persisted version of the rules at each boot
and triggering udev to re-add the devices, this way udev picks up the
persisted rules and names the NICs accordingly.

rhbz#824595

Change-Id: Id432376e4323d4fea4dd2290a9623493e02b5e3f
Signed-off-by: Fabian Deutsch <fabiand@redhat.com>