tests: Allow callables for waiting
Previously it was only possible to wait for a specific time before it
was check if a specific keyword was on the screen, now it is possible to
pass a callable which can do something more fancy (e.g. wait for the
moment when a specififc keyword appears on the screen).

Change-Id: I442f93ec56f929e8f8517132bf54d5c2b1932c60
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>