tests: Ensure that a reboot happened
Previously the reboot testcase just "pinged" the server to tell that the
client is alive. This change introduces a "reboot marker" which is set
before a reboot is happened, this marker resides on /tmp and should be
removed after reboot. reboot_completed now checks if the marker is still
around (failure, no reboot) or not (success, marker is removed).

Change-Id: Ib740426fc0d945ead45c899fccbc09694a0bea80
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>