tests: Improve rebooting in common (lib)
Previosuly a testcase just run reboot to reboot the system, but this
could lead to races in the worst case leading to systemd blocking the
reboot.
With this patch a testcase can use the common.request_reboot function to
initiate a reboot. The testing process will be quit and at last
initiating a reboot.

Change-Id: I82af864c2a917d4f09c85a2460a090c0f05c7679
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>