network: Don't modify tagged interfaces nic names
A dict is used internally to map interface names to the corresponding
interface informations.
Previously the VLAN ID of "VLAN interfaces" was removed from a interface
name (e.g. eth0.42 became eth0) which lead to wrong
NIC-informations-mappings, because the informations corresponding to a
VLAN interface (eth0.42) overwrote the informations of the original
interface (eth0). This lead to obvious problems because informations like
e.g. vendor were missing from the vlan interfaces informations.
This patch prevents the name mangling of tagged interfaces, this way the
interface is later ignored when summarizing the informations.

rhbz#837745

Change-Id: If0afd40a7f9c0b541f06722cee50568e746c0336
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>