o-c-setup: Display and write valid network config
Previosuly, in the case of invalid static network configuration settings
(e.g. invalid IP, GW, ...), the type was set to "static" for the runtime
of the TUI and never changed during the runtime of the TUI.
Now the configuration type is updated correct.

This patch additionally addresse the problem that not all invalid fields
were given in the error message.

rhbz#837614

Change-Id: Ia54d2c5842661c9eac28b7436a1e4159a950e80f
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>