tui: Translate bootprotocol 'none' to 'static'
Previosuly the TUI displayed 'none' on the Status page, if a static IP
configuration was used.
Now the correct term 'static' is used. The problem was that the TUI
passed the low-level BOOTPROTO variable from ifcfg-* to the TUI, where
'none' is the correct value for static IP configurations.

rhbz#857846

Change-Id: I1d28b056657e632d5553b5881767acb42511c935
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>